Add VA coverage table generation

+++ Source/Core/DataAccess/DataTransfer/CoverageTableModel.cs
@@ -15,9 +15,9 @@ namespace MilSpace.DataAccess.DataTransfer
         public double ObservObjArea { get; set; }
         public double ExpectedArea { get; set; }
         public double VisibilityArea { get; set; }
-        public int VisibilityPercent { get; set; }
-        public int ToAllExpectedAreaPercent { get; set; }
-        public int ToAllVisibilityAreaPercent { get; set; }
+        public double VisibilityPercent { get; set; }
+        public double ToAllExpectedAreaPercent { get; set; }
+        public double ToAllVisibilityAreaPercent { get; set; }
         public int ObservPointsSeeCount { get; set; }
     }
 }
